# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BtcZone = (typeof BtcZone)[keyof typeof BtcZone];

export const BtcZone = {
  MAX_ACCUMULATION: "MAX_ACCUMULATION",
  AGGRESSIVE_BUY: "AGGRESSIVE_BUY",
  STANDARD_BUY_LOW: "STANDARD_BUY_LOW",
  STANDARD_BUY_HIGH: "STANDARD_BUY_HIGH",
  TAKE_PROFIT: "TAKE_PROFIT",
} as const;

export interface BtcDashboard {
  /** Current BTC/USD price */
  currentPrice: number;
  /** 200-Week Weighted Moving Average */
  wma200w: number;
  /** 20-Week Exponential Moving Average */
  ema20w: number;
  /** 200-Day Simple Moving Average */
  sma200d: number;
  zone: BtcZone;
  /** Human-readable zone label */
  zoneLabel: string;
  /** Color associated with the zone (hex) */
  zoneColor: string;
  /** Recommended action for the current zone */
  actionText: string;
  /** Timestamp of the last data fetch */
  lastUpdated: string;
}

export interface BtcChartPoint {
  date: string;
  price: number | null;
  wma200w?: number | null;
  ema20w?: number | null;
  sma200d?: number | null;
}

export interface BtcChartData {
  points: BtcChartPoint[];
  currentZone: BtcZone;
}

export interface ApiError {
  error: string;
}

export interface BtcBacktestSummary {
  finalValue: number;
  totalInvested: number;
  netProfit: number;
  returnPct: number;
  maxDrawdown: number;
  dcaFinalValue: number;
  dcaTotalInvested: number;
  outperformance: number;
  btcValue: number;
  cashBalance: number;
  numContributions: number;
  numSells: number;
}

export interface BtcBacktestHistoryPoint {
  date: string;
  portfolioValue: number;
  btcValue: number;
  cashBalance: number;
  dcaValue: number;
  price: number;
  zone: string;
}

export interface BtcBacktestTrade {
  date: string;
  type: string;
  zone: string;
  label: string;
  price: number;
  amount: number;
  btcDelta: number;
}

export interface BtcBacktestZoneStat {
  zone: string;
  label: string;
  count: number;
  totalDeployed: number;
  color: string;
}

export interface BtcBacktestResult {
  startDate: string;
  endDate: string;
  baseInstallment: number;
  startingCash: number;
  summary: BtcBacktestSummary;
  history: BtcBacktestHistoryPoint[];
  trades: BtcBacktestTrade[];
  zoneStats: BtcBacktestZoneStat[];
}

export type GetBtcBacktestParams = {
  /**
   * Simulation start date (YYYY-MM-DD)
   */
  startDate: string;
  /**
   * Base bi-monthly contribution in USD
   */
  baseInstallment?: number;
  /**
   * Starting CASH.to balance in USD
   */
  startingCash?: number;
};