# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: btc
    description: Bitcoin data and strategy indicators
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /btc/dashboard:
    get:
      operationId: getBtcDashboard
      tags: [btc]
      summary: Get BTC dashboard data
      description: Returns current BTC price, technical indicators, and current strategy zone
      responses:
        "200":
          description: BTC dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BtcDashboard"
        "500":
          description: Error fetching data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /btc/backtest:
    get:
      operationId: getBtcBacktest
      tags: [btc]
      summary: Run strategy backtest
      description: Simulates the KISBTC zone-based DCA strategy against historical BTC data
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Simulation start date (YYYY-MM-DD)
        - name: baseInstallment
          in: query
          required: false
          schema:
            type: number
            default: 500
          description: Base bi-monthly contribution in USD
        - name: startingCash
          in: query
          required: false
          schema:
            type: number
            default: 0
          description: Starting CASH.to balance in USD
      responses:
        "200":
          description: Backtest result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BtcBacktestResult"
        "400":
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Backtest error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /btc/chart:
    get:
      operationId: getBtcChart
      tags: [btc]
      summary: Get BTC chart data
      description: Returns historical BTC price with overlaid moving averages for charting
      responses:
        "200":
          description: BTC chart data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BtcChartData"
        "500":
          description: Error fetching data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    BtcZone:
      type: string
      enum:
        - MAX_ACCUMULATION
        - AGGRESSIVE_BUY
        - STANDARD_BUY_LOW
        - STANDARD_BUY_HIGH
        - TAKE_PROFIT
    BtcDashboard:
      type: object
      properties:
        currentPrice:
          type: number
          description: Current BTC/USD price
        wma200w:
          type: number
          description: 200-Week Weighted Moving Average
        ema20w:
          type: number
          description: 20-Week Exponential Moving Average
        sma200d:
          type: number
          description: 200-Day Simple Moving Average
        zone:
          $ref: "#/components/schemas/BtcZone"
        zoneLabel:
          type: string
          description: Human-readable zone label
        zoneColor:
          type: string
          description: Color associated with the zone (hex)
        actionText:
          type: string
          description: Recommended action for the current zone
        lastUpdated:
          type: string
          format: date-time
          description: Timestamp of the last data fetch
      required:
        - currentPrice
        - wma200w
        - ema20w
        - sma200d
        - zone
        - zoneLabel
        - zoneColor
        - actionText
        - lastUpdated
    BtcChartPoint:
      type: object
      properties:
        date:
          type: string
          format: date
        price:
          type: number
          nullable: true
        wma200w:
          type: number
          nullable: true
        ema20w:
          type: number
          nullable: true
        sma200d:
          type: number
          nullable: true
      required:
        - date
        - price
    BtcChartData:
      type: object
      properties:
        points:
          type: array
          items:
            $ref: "#/components/schemas/BtcChartPoint"
        currentZone:
          $ref: "#/components/schemas/BtcZone"
      required:
        - points
        - currentZone
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    BtcBacktestSummary:
      type: object
      properties:
        finalValue: { type: number }
        totalInvested: { type: number }
        netProfit: { type: number }
        returnPct: { type: number }
        maxDrawdown: { type: number }
        dcaFinalValue: { type: number }
        dcaTotalInvested: { type: number }
        outperformance: { type: number }
        btcValue: { type: number }
        cashBalance: { type: number }
        numContributions: { type: number }
        numSells: { type: number }
      required:
        - finalValue
        - totalInvested
        - netProfit
        - returnPct
        - maxDrawdown
        - dcaFinalValue
        - dcaTotalInvested
        - outperformance
        - btcValue
        - cashBalance
        - numContributions
        - numSells
    BtcBacktestHistoryPoint:
      type: object
      properties:
        date: { type: string, format: date }
        portfolioValue: { type: number }
        btcValue: { type: number }
        cashBalance: { type: number }
        dcaValue: { type: number }
        price: { type: number }
        zone: { type: string }
      required:
        - date
        - portfolioValue
        - btcValue
        - cashBalance
        - dcaValue
        - price
        - zone
    BtcBacktestTrade:
      type: object
      properties:
        date: { type: string, format: date }
        type: { type: string }
        zone: { type: string }
        label: { type: string }
        price: { type: number }
        amount: { type: number }
        btcDelta: { type: number }
      required:
        - date
        - type
        - zone
        - label
        - price
        - amount
        - btcDelta
    BtcBacktestZoneStat:
      type: object
      properties:
        zone: { type: string }
        label: { type: string }
        count: { type: number }
        totalDeployed: { type: number }
        color: { type: string }
      required:
        - zone
        - label
        - count
        - totalDeployed
        - color
    BtcBacktestResult:
      type: object
      properties:
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        baseInstallment: { type: number }
        startingCash: { type: number }
        summary:
          $ref: "#/components/schemas/BtcBacktestSummary"
        history:
          type: array
          items:
            $ref: "#/components/schemas/BtcBacktestHistoryPoint"
        trades:
          type: array
          items:
            $ref: "#/components/schemas/BtcBacktestTrade"
        zoneStats:
          type: array
          items:
            $ref: "#/components/schemas/BtcBacktestZoneStat"
      required:
        - startDate
        - endDate
        - baseInstallment
        - startingCash
        - summary
        - history
        - trades
        - zoneStats