# lib/api-zod/src/generated/types/btcBacktestHistoryPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BtcBacktestHistoryPoint {
  date: Date;
  portfolioValue: number;
  btcValue: number;
  cashBalance: number;
  dcaValue: number;
  price: number;
  zone: string;
}