# lib/api-zod/src/generated/types/btcBacktestResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BtcBacktestHistoryPoint } from "./btcBacktestHistoryPoint";
import type { BtcBacktestSummary } from "./btcBacktestSummary";
import type { BtcBacktestTrade } from "./btcBacktestTrade";
import type { BtcBacktestZoneStat } from "./btcBacktestZoneStat";

export interface BtcBacktestResult {
  startDate: Date;
  endDate: Date;
  baseInstallment: number;
  startingCash: number;
  summary: BtcBacktestSummary;
  history: BtcBacktestHistoryPoint[];
  trades: BtcBacktestTrade[];
  zoneStats: BtcBacktestZoneStat[];
}