# lib/api-zod/src/generated/types/btcBacktestSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BtcBacktestSummary {
  finalValue: number;
  totalInvested: number;
  netProfit: number;
  returnPct: number;
  maxDrawdown: number;
  dcaFinalValue: number;
  dcaTotalInvested: number;
  outperformance: number;
  btcValue: number;
  cashBalance: number;
  numContributions: number;
  numSells: number;
}