# lib/api-zod/src/generated/types/btcBacktestTrade.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BtcBacktestTrade {
  date: Date;
  type: string;
  zone: string;
  label: string;
  price: number;
  amount: number;
  btcDelta: number;
}