# lib/api-zod/src/generated/types/btcBacktestZoneStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BtcBacktestZoneStat {
  zone: string;
  label: string;
  count: number;
  totalDeployed: number;
  color: string;
}